Backend/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
anyio==4.3.0
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.5.14
faiss-cpu==1.7.4
fastapi==0.110.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
h11==0.14.0
htbuilder==0.6.2
idna==3.6
importlib_metadata==7.0.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langchain==0.0.154
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mdurl==0.1.2
more-itertools==10.2.0
multidict==6.0.5
mypy-extensions==1.0.0
numexpr==2.9.0
numpy==1.26.4
openapi-schema-pydantic==1.2.4
packaging==24.0
pandas==2.2.1
pillow==10.2.0
protobuf==3.20.3
pyarrow==15.0.1
pydantic==1.10.14
pydeck==0.8.1b0
Pygments==2.17.2
Pympler==1.0.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
semver==3.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.28
st-annotated-text==4.0.1
starlette==0.36.3
streamlit==1.18.1
streamlit-camera-input-live==0.2.0
streamlit-embedcode==0.1.2
streamlit-extras==0.1.5
streamlit-keyup==0.2.3
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
uvicorn==0.28.0
validators==0.22.0
watchdog==4.0.0
yarl==1.9.4
zipp==3.18.0
python-multipart
openapi
tiktoken